refactor: extract useAddedToShoppingListAlert to own file in PDP
